+++ lib/maze/servlets/command_servlet.rb
@@ -8,6 +8,7 @@ module Maze
     # Allows clients to queue up "commands", in the form of Ruby hashes, using Maze::Server.commands.add.  GET
     # requests made to the /command endpoint will then respond with each queued command in turn.
     class CommandServlet < BaseServlet
+
       # Serves the next command, if these is one.
       #
       # @param _request [HTTPRequest] The incoming GET request
@@ -20,7 +21,9 @@ module Maze
           response.body = 'No commands to provide'
           response.status = 400
         else
-          response.body = JSON.pretty_generate(commands.current)
+          command = commands.current
+          command[:uuid] = Maze::Server.command_uuid
+          response.body = JSON.pretty_generate(command)
           response.status = 200
           commands.next
         end