+++ lib/maze/servlets/command_servlet.rb
@@ -8,7 +8,6 @@ module Maze
     # Allows clients to queue up "commands", in the form of Ruby hashes, using Maze::Server.commands.add.  GET
     # requests made to the /command endpoint will then respond with each queued command in turn.
     class CommandServlet < BaseServlet
-
       # Serves the next command, if these is one.
       #
       # @param _request [HTTPRequest] The incoming GET request